+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 * Add RTL top file
 * Add FSM file
+
+## [ver. 0.0.2]
+
+* Add led_dec and sseg_dec modules
+* Fixed sawtooth counter incrementation